+++ js/index.js
@@ -10,8 +10,11 @@ import poc from './components/forms/poc'
 import financial from './components/forms/financial'
 import toggler from './components/toggler'
 import NewProject from './components/forms/new_project'
+import Modal from './mixins/modal'
 
 Vue.use(VTooltip)
+
+Vue.mixin(Modal)
 
 const app = new Vue({
   el: '#app-root',
@@ -25,28 +28,11 @@ const app = new Vue({
     financial,
     NewProject
   },
-  methods: {
-    closeModal: function(name) {
-      this.modals[name] = false
-    },
-    openModal: function (name) {
-      this.modals[name] = true
-    }
-  },
-  data: function() {
-    return {
-      modals: {
-        styleguideModal: false,
-        pendingFinancialVerification: false,
-        pendingCCPOApproval: false,
-      }
-    }
-  },
   mounted: function() {
-    const modalOpen = document.querySelector("#modalOpen");
+    const modalOpen = document.querySelector("#modalOpen")
     if (modalOpen) {
-      const modal = modalOpen.getAttribute("data-modal");
-      this.modals[modal] = true;
+      const modal = modalOpen.getAttribute("data-modal")
+      this.openModal(modal)
     }
   },
   delimiters: ['!{', '}']
