+++ js/mixins/modal.js
@@ -1,0 +1,20 @@
+export default {
+  methods: {
+    closeModal: function(name) {
+      this.modals[name] = false
+    },
+    openModal: function (name) {
+      this.modals[name] = true
+    }
+  },
+  data: function() {
+    return {
+      modals: {
+        styleguideModal: false,
+        newProjectConfirmation: false,
+        pendingFinancialVerification: false,
+        pendingCCPOApproval: false,
+      }
+    }
+  }
+}
